Add files via upload
The data consists of cropped images of 200 subjects, 3 images each, each image of size 24x21. The ﬁle data.mat has a variable face of size (24x21x600). The images corresponding to the person labeled n, n = 1, . . . , 200 can be indexed in Matlab as face(:,:,3*n-2), face(:,:,3*n-1) and face(:,:,3*n). The ﬁrst image is a neutral face, the second image has a facial expression, and the third image has illumination variations.